feat: Add proof of equivalence between semantics
I redefined the notion of non-redundancy to a simplified
one that assumes queries conform to the schema. The
idea is that the definition given by J&O also assumes
this fact, and thus can be simplified (the check on
arguments is not required and is obtained via the field merging
predicate).