fix: Change def of is_field_merging_possible
The previous definition would consider valid selections invalid.
The problem was that I was reusing the same type in context for every
collected field, instead of wrapping each with its corresponding type.
This would cause lookups to fail and cause everything to be considered
invalid.

The lookups are done by using the same type with which each field was
wrapped.

The only thing that is a bit annoying is that one needs the auxiliary definitions for pairs...